pyproject.toml
[project]
name = "my-prefect-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "prefect>=3.5.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
